--- Navigation/Navigation/Profile/Copy/Photocell01.swift
@@ -1,52 +1,0 @@
-//
-//  photocell01.swift
-//  Navigation
-//
-//  Created by dmitriy on 27/6/2023.
-//
-/*
-import UIKit
-
-class PhotoCell: UICollectionViewCell {
-    
-    @IBOutlet weak var imageView: UIImageView!
-    
-}
-
-*/
-
-
-/*
-struct PhotoCell {
-    
-    var image: String
-    
-    static func maketPost() -> [PhotoCell] {
-        
-        var post = [PhotoCell]()
-        post.append(PhotoCell(image: "space3_01"))
-        post.append(PhotoCell(image: "space3_03"))
-        post.append(PhotoCell(image: "space3_06"))
-        post.append(PhotoCell(image: "space3_07"))
-        post.append(PhotoCell(image: "blackBear"))
-        post.append(PhotoCell(image: "blackDolphin"))
-        post.append(PhotoCell(image: "blackEagle"))
-        post.append(PhotoCell(image: "blackFox"))
-        post.append(PhotoCell(image: "blackFlamigo"))
-        post.append(PhotoCell(image: "blackGiraffe"))
-        post.append(PhotoCell(image: "blackGorilla"))
-        post.append(PhotoCell(image: "blackHourse"))
-        post.append(PhotoCell(image: "blackLeopard"))
-        post.append(PhotoCell(image: "blackMonkey"))
-        post.append(PhotoCell(image: "blackMoose"))
-        post.append(PhotoCell(image: "blackPanda"))
-        post.append(PhotoCell(image: "blackRaccoon"))
-        post.append(PhotoCell(image: "blackMonkey2"))
-        post.append(PhotoCell(image: "blackShark"))
-        post.append(PhotoCell(image: "blackSnake"))
-        post.append(PhotoCell(image: "blackTiger"))
-    
-        return post
-    }
-}
-*/
