--- Navigation/Navigation/Profile/Copy/PhotosViewController01.swift
@@ -1,61 +1,0 @@
-//
-//  PhotosViewController01.swift
-//  Navigation
-//
-//  Created by dmitriy on 27/6/2023.
-//
-/*
-import UIKit
-class PhotosViewController: UIViewController {
-
-    let cellIdentifier = "photoCell"
-    let sectionInsets = UIEdgeInsets(top: 10, left: 10, bottom: 10, right: 10)
-    let itemsPerRow: CGFloat = 3
-    
-    var collectionView: UICollectionView!
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setupCollectionView()
-        collectionView.register(UINib(nibName: "PhotoCell", bundle: nil), forCellWithReuseIdentifier: cellIdentifier)
-    }
-    
-    func setupCollectionView() {
-        let layout = UICollectionViewFlowLayout()
-        layout.minimumInteritemSpacing = sectionInsets.left
-        layout.minimumLineSpacing = sectionInsets.left
-        layout.sectionInset = sectionInsets
-        let itemWidth = (view.frame.width - sectionInsets.left - sectionInsets.right - (itemsPerRow - 1) * layout.minimumInteritemSpacing) / itemsPerRow
-        layout.itemSize = CGSize(width: itemWidth, height: itemWidth)
-        
-        collectionView = UICollectionView(frame: view.bounds, collectionViewLayout: layout)
-        collectionView.backgroundColor = .white
-        collectionView.dataSource = self
-        collectionView.delegate = self
-        collectionView.register(UICollectionViewCell.self, forCellWithReuseIdentifier: cellIdentifier)
-        collectionView.contentInsetAdjustmentBehavior = .always
-        collectionView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: 20, right: 0)
-
-        view.addSubview(collectionView)
-    }
-
-}
-
-extension PhotosViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 9
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellIdentifier, for: indexPath) as! PhotoCell
-        cell.imageView.image = UIImage(named: "space3_01\(indexPath.item + 1)")
-        return cell
-    }
-}
-
-extension PhotosViewController: UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        
-    }
-}
-*/
