+++ Navigation/Navigation/Profile/PhotosHorizontalCollectionViewCell.swift
@@ -7,22 +7,18 @@
 
 import UIKit
 
-final class PhotosSlideCollectionViewCell: UICollectionViewCell {
-    
-    //MARK: - Property
-    
+ class PhotosHorizontalCollectionViewCell: UICollectionViewCell {
+
     private lazy var postImageView: UIImageView = {
         let imageView = UIImageView()
         imageView.translatesAutoresizingMaskIntoConstraints = false
         imageView.contentMode = .scaleAspectFit
+        
         return imageView
     }()
-        
-    //MARK: - Inits
     
     override init(frame: CGRect) {
         super.init(frame: frame)
-        custmizeCell()
         addSubviews()
         setupContraints()
     }
@@ -30,24 +26,17 @@ final class PhotosSlideCollectionViewCell: UICollectionViewCell {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
-    //MARK: - Functions
-    
-    private func custmizeCell() {
-    }
-    
-    func setupCell(model: PostImages) {
+     
+    func setupCell(model: ImagesHeader) {
         postImageView.image = UIImage(named: model.image)
     }
     
-    private func addSubviews(){
+    func addSubviews(){
         contentView.addSubview(postImageView)
        
     }
     
-    //MARK: - Constraints
-    
-    private func setupContraints() {
+     func setupContraints() {
         
         NSLayoutConstraint.activate([
             postImageView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 12),
