+++ Navigation/Navigation/Profile/PhotosTableViewCell.swift
@@ -1,5 +1,5 @@
 //
-//  PhotosTableViewCell.swift
+//  PhotosController.swift
 //  Navigation
 //
 //  Created by dmitriy on 26/6/2023.
@@ -8,154 +8,126 @@
 import UIKit
 
 class PhotosTableViewCell: UITableViewCell {
-    
-    //MARK: - Properties
-    
-    var imageView1: UIImageView!
-    var imageView2: UIImageView!
-    var imageView3: UIImageView!
-    var imageView4: UIImageView!
-    
-    private let contentWhiteView: UIView = {
-        let view = UIView()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        return view
-    }()
-    
-    private let imageArrow: UIImageView = {
-        let imageView = UIImageView()
-        imageView.image =  UIImage(systemName: "arrow.right")
-        imageView.contentMode = .scaleAspectFill
-        imageView.tintColor = .black
-        imageView.clipsToBounds = true
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        return imageView
-    }()
+
+    var tapHandler: (() -> Void)? = nil
+        
+    private let postImages = ImagesHeader.maketPost()
     
     private let headerLabel: UILabel = {
-        let label = UILabel()
-        label.font = UIFont.boldSystemFont(ofSize: 24)
-        label.text = "Photos"
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
+        let headerLabel = UILabel()
+        headerLabel.font = UIFont.boldSystemFont(ofSize: 24)
+        headerLabel.text = "Photos"
+        headerLabel.translatesAutoresizingMaskIntoConstraints = false
+        
+        return headerLabel
+    }()
+    
+    private lazy var imageArrow: UIImageView = {
+        let imageArrow = UIImageView()
+        imageArrow.image =  UIImage(systemName: "arrow.right")
+        imageArrow.contentMode = .scaleAspectFill
+        imageArrow.tintColor = .black
+        imageArrow.clipsToBounds = true
+        imageArrow.translatesAutoresizingMaskIntoConstraints = false
+        
+        return imageArrow
+    }()
+    
+    private lazy var collectionView: UICollectionView = {
+        let layout = UICollectionViewFlowLayout()
+        layout.scrollDirection = .horizontal
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
+        collectionView.translatesAutoresizingMaskIntoConstraints = false
+        collectionView.backgroundColor = .systemGray2
+        collectionView.dataSource = self
+        collectionView.register(PhotosHorizontalCollectionViewCell.self, forCellWithReuseIdentifier: PhotosHorizontalCollectionViewCell.identifier)
+        collectionView.delegate = self
+        collectionView.backgroundColor = .white
+        collectionView.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(tapAction)))
+        
+        return collectionView
+    }()
+    
+    @objc func tapAction() {
+        if let action = self.tapHandler { action() }
+    }
+    
+    private let postImageView: UIImageView = {
+        let postImageView = UIImageView()
+        postImageView.translatesAutoresizingMaskIntoConstraints = false
+        postImageView.contentMode = .scaleAspectFit
+        return postImageView
     }()
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
-        addImage()
         addSubviews()
         setupContraints()
     }
+    
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    //MARK: - Functions
-    
-    //    override func prepareForReuse() {
-    //        super.prepareForReuse()
-    //        imageView1.image = nil
-    //        imageView2.image = nil
-    //        imageView3.image = nil
-    //        imageView4.image = nil
-    //        headerLabel.text = nil
-    //        imageArrow.image = nil
-    //    }
-    
-    private func addImage() {
-        imageView1 = createImageView(nameImageView: "imageView1", nameImage: "blackCat")
-        imageView2 = createImageView(nameImageView: "imageView2", nameImage: "blackDog")
-        imageView3 = createImageView(nameImageView: "imageView3", nameImage: "blackMouse")
-        imageView4 = createImageView(nameImageView: "imageView4", nameImage: "blackElefant")
+    func setupCell(model: ImagesHeader) {
+        postImageView.image = UIImage(named: model.image)
     }
     
     private func addSubviews(){
-        
-        contentView.addSubview(contentWhiteView)
-        contentView.addSubview(imageView1)
-        contentView.addSubview(imageView2)
-        contentView.addSubview(imageView3)
-        contentView.addSubview(imageView4)
-        contentView.addSubview(headerLabel)
+        contentView.addSubview(collectionView)
         contentView.addSubview(imageArrow)
-        
+        contentView.addSubview(headerLabel)
     }
-    
-    //MARK: - Constraints
     
     private func setupContraints() {
-
-        let imageSize: CGFloat = contentView.frame.width / 4
-            
+        
         NSLayoutConstraint.activate([
             
-            //             contentWhiteView
-            contentWhiteView.topAnchor.constraint(equalTo: contentView.topAnchor),
-            contentWhiteView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-            contentWhiteView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
-            contentWhiteView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            collectionView.topAnchor.constraint(equalTo: contentView.topAnchor),
+            collectionView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            collectionView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
+            collectionView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
             
-            //            headerLabel
-            headerLabel.topAnchor.constraint(equalTo: contentWhiteView.topAnchor ,constant: Metric.labelInset),
-            headerLabel.leadingAnchor.constraint(equalTo: contentWhiteView.leadingAnchor,constant: Metric.labelInset),
-            
-            //            imageView1
-            imageView1.widthAnchor.constraint(equalToConstant: imageSize),
-            imageView1.heightAnchor.constraint(equalToConstant: imageSize),
-            imageView1.topAnchor.constraint(equalTo: headerLabel.bottomAnchor,constant: Metric.imageInset),
-            imageView1.leadingAnchor.constraint(equalTo: contentWhiteView.leadingAnchor,constant: Metric.imageInset),
-            imageView1.bottomAnchor.constraint(equalTo: contentWhiteView.bottomAnchor, constant: -Metric.imageInset),
-            
-            //            imageView2
-            imageView2.widthAnchor.constraint(equalToConstant: imageSize),
-            imageView2.heightAnchor.constraint(equalToConstant: imageSize),
-            imageView2.topAnchor.constraint(equalTo: headerLabel.bottomAnchor,constant: Metric.imageInset),
-            imageView2.leadingAnchor.constraint(equalTo: imageView1.trailingAnchor,constant: Metric.imageLeftInset),
-            imageView2.bottomAnchor.constraint(equalTo: contentWhiteView.bottomAnchor, constant: -Metric.imageInset),
-            
-            //            imageView3
-            imageView3.widthAnchor.constraint(equalToConstant: imageSize),
-            imageView3.heightAnchor.constraint(equalToConstant: imageSize),
-            imageView3.topAnchor.constraint(equalTo: headerLabel.bottomAnchor,constant: Metric.imageInset),
-            imageView3.leadingAnchor.constraint(equalTo: imageView2.trailingAnchor,constant: Metric.imageLeftInset),
-            imageView3.bottomAnchor.constraint(equalTo: contentWhiteView.bottomAnchor, constant: -Metric.imageInset),
-            
-            //            imageView4
-            imageView4.widthAnchor.constraint(equalToConstant: imageSize),
-            imageView4.heightAnchor.constraint(equalToConstant: imageSize),
-            imageView4.topAnchor.constraint(equalTo: headerLabel.bottomAnchor,constant: Metric.imageInset),
-            imageView4.leadingAnchor.constraint(equalTo: imageView3.trailingAnchor,constant: Metric.imageLeftInset),
-            imageView4.bottomAnchor.constraint(equalTo: contentWhiteView.bottomAnchor, constant: -Metric.imageInset),
-            
-            //            ImageArrow
             imageArrow.widthAnchor.constraint(equalToConstant: Metric.imageArrowInset),
-            imageArrow.trailingAnchor.constraint(equalTo: contentWhiteView.trailingAnchor, constant: -12),
-            imageArrow.centerYAnchor.constraint(equalTo: headerLabel.centerYAnchor)
+            imageArrow.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -12),
+            imageArrow.centerYAnchor.constraint(equalTo: headerLabel.centerYAnchor),
+            
+            headerLabel.topAnchor.constraint(equalTo: topAnchor ,constant: Metric.labelInset),
+            headerLabel.leadingAnchor.constraint(equalTo: leadingAnchor,constant: Metric.labelInset)
         ])
     }
-    
-    //MARK: - Action
-    
-    let photosViewController =  {
-        PhotosViewController()
-    }()
 }
 
-extension PhotosTableViewCell {
-    func createImageView(nameImageView: String, nameImage: String) -> UIImageView{
-        
-        let nameImageView: UIImageView = {
-            let imageView = UIImageView()
-            imageView.layer.borderColor = UIColor.white.cgColor
-            imageView.layer.borderWidth = 3
-            imageView.layer.cornerRadius = 6
-            imageView.image = UIImage(named: nameImage)
-            imageView.contentMode = .scaleAspectFill
-            imageView.clipsToBounds = true
-            imageView.translatesAutoresizingMaskIntoConstraints = false
-            return imageView
-        }()
-        return nameImageView
+
+extension PhotosTableViewCell: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        postImages.count
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: PhotosHorizontalCollectionViewCell.identifier, for: indexPath) as! PhotosHorizontalCollectionViewCell
+        cell.setupCell(model: postImages[indexPath.row])
+        return cell
+    }
+}
+
+extension PhotosTableViewCell: UICollectionViewDelegateFlowLayout {
+    private var inset: CGFloat { return 8}
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let width = (collectionView.bounds.width - 10 * 4) / 4
+        return CGSize(width: width, height: width)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
+        inset
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
+        UIEdgeInsets(top: inset, left: inset, bottom: inset, right: inset)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
+        inset
     }
 }
 
@@ -166,4 +138,8 @@ extension PhotosTableViewCell {
         static let imageLeftInset: CGFloat = 8
         static let imageArrowInset: CGFloat = 30
     }
+}
+
+extension PhotosTableViewCell {
+    
 }
